requirements.txt
gunicorn==20.1.0
uvicorn==0.16.0
fastapi==0.79.0
numpy==1.23.1
tensorflow-cpu
keras